pyproject.toml
[project]
name = "euler-copilot-rag"
version = "0.9.5"
description = "EulerCopilot 后端服务"
requires-python = "==3.11.6"
dependencies = [
    "aiofiles==24.1.0",
    "aiomysql==0.2.0",
    "apscheduler==3.10.0",
    "asgi-correlation-id==4.3.1",
    "asyncpg==0.29.0",
    "beautifulsoup4==4.12.3",
    "chardet==5.2.0",
    "cryptography==42.0.2",
    "fastapi==0.110.2",
    "fastapi-pagination==0.12.19",
    "httpx==0.27.0",
    "itsdangerous==2.1.2",
    "jieba==0.42.1",
    "langchain==0.3.7",
    "langchain-openai==0.2.5",
    "minio==7.2.4",
    "more-itertools==10.1.0",
    "numpy==1.26.4",
    "openai==1.53.0",
    "opencv-python==4.9.0.80",
    "openpyxl==3.1.2",
    "paddleocr==2.9.1",
    "paddlepaddle==2.6.2",
    "pandas==2.2.2",
    "pgvector==0.2.5",
    "pillow==10.3.0",
    "psycopg2-binary==2.9.9",
    "pydantic==2.7.4",
    "pymupdf==1.24.13",
    "python-docx==1.1.0",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.9",
    "python-pptx==0.6.23",
    "pytz==2020.5",
    "pyyaml==6.0.1",
    "redis==5.0.3",
    "requests==2.32.2",
    "scikit-learn==1.5.0",
    "sqlalchemy==2.0.23",
    "starlette==0.37.2",
    "tika==2.6.0",
    "tiktoken==0.8.0",
    "urllib3==2.2.1",
    "uvicorn==0.21.0",
    "xlrd==2.0.1",
]

[[tool.uv.index]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
